pyproject.toml
[build-system]
requires = [
    "scikit-build-core>=0.10",
    "nanobind>=2.0.0",
    "pybind11-stubgen",
    "setuptools-scm>=8",
]
build-backend = "scikit_build_core.build"

[project]
name = "rapids_singlecell"
description = "running single cell analysis on Nvidia GPUs"
requires-python = ">=3.12, <3.14"
license = { file = "LICENSE" }
authors = [ { name = "Severin Dicks" } ]
readme = { file = "README.md", content-type = "text/markdown" }
dynamic = [ "version" ]

dependencies = [
    "anndata>=0.10.0",
    "scanpy>=1.10.0",
    "numpy>=1.17.0",
    "scipy>=1.4",
    "pandas",
    "natsort",
    "scikit-misc>=0.1.3",
    "scikit-learn",
    "matplotlib>=3.4",
    "tqdm",
    "statsmodels>=0.12.0",
    "scikit-image",
    "docrep",
]

[project.optional-dependencies]
rapids13 = [ "cupy-cuda13x", "cudf-cu13>=25.10", "cuml-cu13>=25.10", "cugraph-cu13>=25.10", "cuvs-cu13>=25.10" ]
rapids12 = [ "cupy-cuda12x", "cudf-cu12>=25.10", "cuml-cu12>=25.10", "cugraph-cu12>=25.10", "cuvs-cu12>=25.10" ]

doc = [
    "sphinx>=4.5.0",
    "sphinx-copybutton",
    "nbsphinx>=0.8.12",
    "myst-nb",
    "sphinx-design",
    "sphinx-tabs",
    "sphinxext-opengraph",
    "sphinxcontrib.bibtex",
    "scanpydoc[typehints,theme]>=0.9.4",
    "dask",
    "pytest",
]
test-minimal = [
    "pytest",
    "pytest-cov",
    "pytest-timeout",
    "pytest-rerunfailures",
    "anyio",
    "pooch",
    "profimp",
    "scanpy>=1.10.0",
    "bbknn",
    "decoupler",
    "fast-array-utils",
]
test = [
    "rapids_singlecell[test-minimal]",
    "igraph",
]

dev = [
    "hatch-vcs",
    "pre-commit",
]

[project.urls]
Documentation = "https://rapids-singlecell.readthedocs.io"
Source = "https://github.com/scverse/rapids_singlecell"

[tool.ruff]
src = [ "src" ]
line-length = 88
lint.select = [
    "F",       # Errors detected by Pyflakes
    "E",       # Error detected by Pycodestyle
    "W",       # Warning detected by Pycodestyle
    "I",       # isort
    "TID",     # flake8-tidy-imports
    "C4",      # flake8-comprehensions
    "BLE",     # flake8-blind-except
    "FBT",     # No boolean positional parameters
    "UP",      # pyupgrade
    "TID251",  # Banned imports
    "RUF100",  # Report unused noqa directives
    "ICN",     # Follow import conventions
    "PTH",     # Pathlib instead of os.path
    "TC",      # manage type checking blocks
    "PLR0917", # Ban APIs with too many positional parameters
]

lint.ignore = [
    # line too long -> we accept long comment lines; formatting gets rid of long code lines
    "E501",
    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
    "E731",
    # allow I, O, l as variable names -> I is the identity matrix
    "E741",
    # Missing docstring in public package
    "F403",
]

[tool.ruff.lint.per-file-ignores]
"docs/*" = [ "I" ]
"tests/*" = [ "D", "FBT" ]
"*/__init__.py" = [ "F401" ]
[tool.ruff.lint.isort]
known-first-party = [ "rapids_singlecell" ]
required-imports = [ "from __future__ import annotations" ]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--strict-markers",
    "--timeout=120",
]
markers = [
    "gpu: tests that use a GPU (currently unused, but needs to be specified here as we import anndata.tests.helpers, which uses it)",
]

[tool.setuptools_scm]
write_to = "src/rapids_singlecell/_version.py"
# Optional but useful:
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"

[tool.scikit-build]
# Use limited ABI wheels (one wheel for all Python minor versions on one platform)
wheel.py-api = "cp312"
wheel.packages = [ "src/rapids_singlecell", "src/testing" ]
cmake.version = ">=3.24"
cmake.build-type = "Release"
ninja.version = ">=1.10"
experimental = false
# Default: use native arch for source builds (auto-detect user's GPU)
# CI overrides this via SKBUILD_CMAKE_ARGS for precompiled wheels
cmake.args = [ "-DCMAKE_CUDA_ARCHITECTURES=native" ]
build-dir = "build"
metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
sdist.include = [ "src/rapids_singlecell/_version.py" ]

# Use abi3audit to catch issues with Limited API wheels
[tool.cibuildwheel.linux]
repair-wheel-command = [
    "auditwheel repair -w {dest_dir} {wheel}",
    "pipx run abi3audit --strict --report {wheel}",
]
[tool.cibuildwheel.macos]
repair-wheel-command = [
    "delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}",
    "pipx run abi3audit --strict --report {wheel}",
]
[tool.cibuildwheel.windows]
repair-wheel-command = [
    "copy {wheel} {dest_dir}",
    "pipx run abi3audit --strict --report {wheel}",
]

[tool.codespell]
skip = '*.ipynb,*.csv'
ignore-words-list = "nd"

[tool.coverage.run]
data_file = "test-data/coverage"
source_pkgs = [ "rapids_singlecell" ]
omit = [ "tests/*", "src/testing/*" ]

[tool.coverage.xml]
output = "test-data/coverage.xml"

[tool.coverage.paths]
source = [ "./src", "**/site-packages" ]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
]
